HUHU

+++ src/main/java/com/driver/SavingsAccount.java
@@ -25,7 +25,7 @@ public class SavingsAccount extends BankAccount{
 
     public double getSimpleInterest(int years){
         // Return the final amount considering that bank gives simple interest on current amount
-        return (double) getBalance() * (1 + (rate * years)/100);
+        return  getBalance() * (1 + (rate * years)/100);
     }
 
     public double getCompoundInterest(int times, int years){
